+++ tests/test_workflow_template_service.py
@@ -3,9 +3,9 @@ from unittest.mock import Mock
 from argo_workflows.api_client import ApiClient
 from argo_workflows.apis import WorkflowTemplateServiceApi
 from argo_workflows.models import (
+    IoArgoprojWorkflowV1alpha1WorkflowSpec,
     IoArgoprojWorkflowV1alpha1WorkflowTemplate,
     IoArgoprojWorkflowV1alpha1WorkflowTemplateCreateRequest,
-    IoArgoprojWorkflowV1alpha1WorkflowTemplateSpec,
     ObjectMeta,
 )
 
@@ -29,7 +29,7 @@ def test_ws_calls_create_as_expected():
     ws = WorkflowTemplateService(host='https://abc.com', token='abc')
     ws.service = mock_service
     t = IoArgoprojWorkflowV1alpha1WorkflowTemplate(
-        metadata=ObjectMeta(), spec=IoArgoprojWorkflowV1alpha1WorkflowTemplateSpec()
+        metadata=ObjectMeta(), spec=IoArgoprojWorkflowV1alpha1WorkflowSpec()
     )
     ws.create(t)